+++ src/index.ts
@@ -68,6 +68,7 @@ export * from './components/form';
 export * from './components/grid';
 export * from './components/header';
 export * from './components/highlight';
+export * from './components/hover-card';
 export * from './components/icon';
 export * from './components/icon-button';
 export * from './components/image';